requirements.txt
pyTelegramBotAPI
loguru
python-daemon